ASoC: wm_adsp: Don't init cache from DSP memory if control is write-only
For controls marked write-only don't initialize the cache from the
content of the DSP memory. We stil need the cache for any new data
that is written to this control, and we need to return something
for a read of the ALSA control because most user-side code assumes
all ALSA controls are readable. The cache is already created zero-
filled so the only change needed is to skip populating it from
DSP memory if the control isn't readable.

Change-Id: I19639a1fba87fec5a2a359cb6e1251d54bd11641
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>